Add console wallet passphrase encryption
- When the console wallet starts up, check if per field encryption has
  been applied. If not, prompt the user for a password and confirmation,
  and then encrypt wallet.
- Allow the password to be provided by prompt, command line argument,
  config file, or environment variable (TARI_WALLET_PASSWORD) for
  automation purposes.
- Change the wallet password by passing `--change-password`
- Refactor console wallet initialization functions from `main.rs` to
  their own `init` module to clean up main.
- Remove emojis from base node status element.
- Fix bug where setting a custom base node did not update the base node
  detail element
