+++ applications/tari_console_wallet/Cargo.toml
@@ -21,13 +21,13 @@ futures = { version = "^0.3.1", default-features = false, features = ["alloc"]}
 crossterm = { version = "0.17"}
 rand = "0.7.2"
 unicode-width = "0.1"
-lazy_static = "1.4.0"
 unicode-segmentation = "1.6.0"
 log = { version = "0.4.8", features = ["std"] }
 structopt = { version = "0.3.13", default_features = false }
 config = { version = "0.9.3" }
 qrcode = { version = "0.12" }
 dirs-next = "1.0.2"
+rpassword = "5.0"
 strum = "^0.19"
 strum_macros = "^0.19"
 time = "0.1.38"
