+++ common/src/configuration/bootstrap.rs
@@ -113,8 +113,15 @@ pub struct ConfigBootstrap {
     /// This will clean out the orphans db at startup
     #[structopt(long, alias("clean_orphans_db"))]
     pub clean_orphans_db: bool,
+    /// Enable Mining
     #[structopt(long, alias("enable_mining"))]
     pub enable_mining: bool,
+    /// Supply the password for the console wallet
+    #[structopt(long)]
+    pub password: Option<String>,
+    /// Change the password for the console wallet
+    #[structopt(long, alias("update-password"))]
+    pub change_password: bool,
 }
 
 impl Default for ConfigBootstrap {
@@ -131,6 +138,8 @@ impl Default for ConfigBootstrap {
             command: None,
             clean_orphans_db: false,
             enable_mining: false,
+            password: None,
+            change_password: false,
         }
     }
 }
